+++ 每日单词/dir.json
@@ -1,0 +1,19 @@
+[
+ "7.26.json",
+ "7.27.json",
+ "7.28.json",
+ "7.29.json",
+ "7.30.json",
+ "7.31.json",
+ "8.1.json",
+ "8.10.json",
+ "8.11.json",
+ "8.2.json",
+ "8.3.json",
+ "8.4.json",
+ "8.5.json",
+ "8.6.json",
+ "8.7.json",
+ "8.8.json",
+ "8.9.json"
+]
