--- 每日单词/dir.py
@@ -1,15 +1,0 @@
-'''每日单词目录
-Version : 1.0.0.0
-Date : 2024/02/10 17:45
-Author : Long17369
-'''
-
-import json         # 用于读取json
-
-class Dirtory():
-    def __init__(self) -> None:
-        pass
-    
-    def getDirtory():
-        
-        ...
